requirements.txt
pytest
pytest-cov
python-dotenv
openai
